Update contentScript.js
I  used this extension and it was having a problem. So I made some changes.
Instead of waiting for 3 seconds to save and next button(assuming that next question will load under 3seconds). this method sometime fails when question takes more than 3 seconds to load, I added a interval of 1 seconds, in every one second it will check whether the next question is loaded, and then it will click.